+++ DesignToolbox/DesignToolbox/Pages/Components/ColoredSurface/NamedColorMode.swift
@@ -12,6 +12,7 @@
 //
 
 import OUDS
+import OUDSComponents
 import OUDSTokensSemantic
 
 enum NamedColorMode: CaseIterable, CustomStringConvertible {